UserManager for RaptorUser
Refactored code involving creating users into new RaptorUserManager
